Optimised FASTAPI code
